Connection-Drop-Proof Run-Once Data Trasfer Script
Connection drops during data transfer can be a nuisance when many files (and/or or much data) have to be transmitted/received. Unfortunately, SciNet will not let me log in without a password and I am too lazy to monitor file transfers overnight (A graduate student needs sleep!). This script allows me to transfer files between my current machine and SciNet in a run-once setting. The best part is: I do not have to store my password in a file! I just have to provide it at the beginning of script execution and it stores it in its working memory for the duration of its run. Then, it forgets. =)

This can be useful when I need large files, or a large number of files, transferred: I once had to transfer ~30,000 flat files, amounting to about 32 GB. [The other option could have been to go to the System Administrator at SciNet and ask for help, but I am not a big fan of the bureaucracy this might have involved. I guess I have no idea if SciNet is run by real people